[chg] Removed most of the locks in FanControl. [chg] Made all public methods of FanControl thread safe.

+++ Libraries/StagWare.FanControl.Service/FanControlService.cs
@@ -52,7 +52,7 @@ namespace StagWare.FanControl.Service
 
         #region IFanControlService implementation
 
-        public void SetTargetFanSpeed(double value, int fanIndex)
+        public void SetTargetFanSpeed(float value, int fanIndex)
         {
             if (fanControl != null)
             {
